requirements.txt
discord.py==2.3.2
mcrcon==0.7.0
